requirements.txt
matplotlib
numpy
pandas<2.0.0
tensorboard
umap-learn
hdbscan
bertopic<0.9.0
scikit-learn<1.2
jupyter
hyperopt<0.2.6
torch
transformers
datasets >= 1.8.0, < 1.10.0
accelerate
